refactor(precache): move pre cache function to a utils file
